python-worker/requirements.txt
redis==5.0.1
psycopg2-binary==2.9.9
opencv-python-headless==4.8.1.78
requests==2.31.0
numpy<2.0.0

